webapp/requirements.txt
Django==1.4.2

#psycopg2==2.4.5

#search engine
#Whoosh==2.4.1
#mmseg==1.3.0

#tag cloud
#django-taggit
#django-taggit-templatetags
#django-templatetag-sugar

#account register
django-userena
easy_thumbnails
PIL
django-simple-captcha

#permission
django-guardian

#tools
yolk

#migrate
South

#manage js and css
django-sekizai

#Rest
djangorestframework
markdown

